Tidying things up a bit
