+++ pages/about/about.html
@@ -10,49 +10,21 @@
 			integrity="sha384-4bw+/aepP/YC94hEpVNVgiZdgIC5+VKNBQNGCHeKRQN+PtmoHDEXuppvnDJzQIu9"
 			crossorigin="anonymous"
 		/>
-
 		<link rel="stylesheet" href="https://unpkg.com/aos@next/dist/aos.css" />
-		<link rel="stylesheet" href="../about/about.css" />
-		<link
-			rel="font-family: 'Borel', cursive;
-	font-family: 'Inter', sans-serif;
-	font-family: 'Quicksand', sans-serif;
-	font-family: 'Raleway', sans-serif;"
-		/>
+		<link rel="stylesheet" href="/index.css" />
+		<link rel="stylesheet" href="./about.css" />
+		<script
+			src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/js/bootstrap.bundle.min.js"
+			integrity="sha384-HwwvtgBNo3bZJJLYd8oVXjrBZt8cqVSpeBNS5n7C8IVInixGAoxmnlMuBnhbgrkm"
+			crossorigin="anonymous"
+			defer
+		></script>
+		<script type="module" src="./about.js"></script>
 	</head>
 
 	<body>
 		<header>
-			<!-- navbar -->
-			<nav class="navbar navbar-expand-lg bg-body-tertiary">
-				<div class="container-fluid">
-					<a class="navbar-brand" href="/pages/homepage/homepage.html">
-						<img src="/assets/img/logo.png" alt="Logo" width="50" height="50" class="rounded-circle align-text-" />
-						Beer Craft Collective</a
-					>
-					<button
-						class="navbar-toggler"
-						type="button"
-						data-bs-toggle="collapse"
-						data-bs-target="#navbarSupportedContent"
-						aria-controls="navbarSupportedContent"
-						aria-expanded="false"
-						aria-label="Toggle navigation"
-					>
-						<span class="navbar-toggler-icon"></span>
-					</button>
-					<div class="collapse navbar-collapse" id="navbarSupportedContent">
-						<div class="navbar-nav mb-1 mx-3">
-							<a href="/pages/homepage/homepage.html" aria-current="page" class="nav-item nav-link mx-5">Inicio</a>
-							<a href="/pages/about/about.html" class="nav-item nav-link active mx-5">Nosotros</a>
-							<a href="/pages/contact/contact.html" class="nav-item nav-link mx-5">Contacto</a>
-							<a href="/pages/cart/cart.html" class="nav-item nav-link mx-5">Carrito</a>
-						</div>
-					</div>
-				</div>
-			</nav>
-
-			<!-- finaliza navbar -->
+			<nav></nav>
 		</header>
 		<main class="container text-center mt-2">
 			<div class="container">
@@ -281,16 +253,7 @@
 		</main>
 
 		<footer></footer>
-		<script
-			src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/js/bootstrap.bundle.min.js"
-			integrity="sha384-HwwvtgBNo3bZJJLYd8oVXjrBZt8cqVSpeBNS5n7C8IVInixGAoxmnlMuBnhbgrkm"
-			crossorigin="anonymous"
-		></script>
-
-		<script type="module" src="./about.js"></script>
-		<script src="../../components/footer.js" type="module"></script>
-		<!-- 
-		script animaciones aos -->
+		<!-- script animaciones aos -->
 		<script src="https://unpkg.com/aos@next/dist/aos.js"></script>
 		<script>
 			AOS.init({ once: true, disable: 'mobile' });
