+++ pages/contact/contact.html
@@ -10,21 +10,15 @@
 			integrity="sha384-4bw+/aepP/YC94hEpVNVgiZdgIC5+VKNBQNGCHeKRQN+PtmoHDEXuppvnDJzQIu9"
 			crossorigin="anonymous"
 		/>
-		<link
-			rel="font-family: 'Borel', cursive;
-			font-family: 'Inter', sans-serif;
-			font-family: 'Quicksand', sans-serif;
-			font-family: 'Raleway', sans-serif;"
-		/>
 		<script
 			src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/js/bootstrap.bundle.min.js"
 			integrity="sha384-HwwvtgBNo3bZJJLYd8oVXjrBZt8cqVSpeBNS5n7C8IVInixGAoxmnlMuBnhbgrkm"
 			crossorigin="anonymous"
 			defer
 		></script>
-		<link rel="stylesheet" href="contact.css" />
+		<link rel="stylesheet" href="/index.css" />
+		<link rel="stylesheet" href="./contact.css" />
 		<script type="module" src="./contact.js"></script>
-		<script type="module" src="/components/footer.js"></script>
 	</head>
 
 	<body>
@@ -33,21 +27,42 @@
 		<main class="container text-center my-5">
 			<div class="container d-flex justify-content-center">
 				<div class="row col-7 col-lg-4">
-					<form id="contact-form" class="text-center" action="https://formsubmit.co/5f0f9995143bfb7bcf4d49ef2a9749b1" method="POST">
+					<form
+						id="contact-form"
+						class="text-center"
+						action="https://formsubmit.co/5f0f9995143bfb7bcf4d49ef2a9749b1"
+						method="POST"
+					>
 						<div class="mb-4">
-						  <label for="name" class="form-label">Nombre</label>
-						  <input value="" name="name" type="text" class="form-control border-dark-subtle" id="name" maxlength="50">
+							<label for="name" class="form-label">Nombre</label>
+							<input
+								value=""
+								name="name"
+								type="text"
+								class="form-control border-dark-subtle"
+								id="name"
+								maxlength="50"
+							/>
 						</div>
 						<div class="mb-4">
-						  <label for="email" class="form-label">Email</label>
-						  <input value="" name="email" type="email" class="form-control border-dark-subtle" id="email" maxlength="256">
+							<label for="email" class="form-label">Email</label>
+							<input
+								value=""
+								name="email"
+								type="email"
+								class="form-control border-dark-subtle"
+								id="email"
+								maxlength="256"
+							/>
 						</div>
 						<div class="mb-4">
 							<label for="message" class="form-label">Tu mensaje</label>
-							<textarea class="form-control border-dark-subtle" name="message" id="message" maxlength="500">lo que sea</textarea>
-						 </div>
+							<textarea class="form-control border-dark-subtle" name="message" id="message" maxlength="500">
+lo que sea</textarea
+							>
+						</div>
 						<button type="submit" id="submit-btn" class="btn btn-warning rounded-pill">Enviar</button>
-					 </form>
+					</form>
 				</div>
 			</div>
 		</main>
