+++ pages/products/products.html
@@ -19,12 +19,10 @@
 		<link rel="stylesheet" href="/index.css" />
 		<link rel="stylesheet" href="./products.css" />
 		<script type="module" src="./products.js"></script>
-		<script type="module" src="./productAdd.js"></script>
 	</head>
 	<body>
 		<nav></nav>
 		<main class="container">
-			<!--formulario para agregar producto-->
 			<section class="container my-3">
 				<h3 class="d-flex justify-content-center">Agregar Producto</h3>
 				<form id="form-product" class="mx-auto row border border-dark-subtlek w-75 d-flex">
