+++ readme.md
@@ -1,1 +1,8 @@
 Reminder for handling having the products on same page as the start page! (or not)
+
+<link
+	rel="font-family: 'Borel', cursive;
+	font-family: 'Inter', sans-serif;
+	font-family: 'Quicksand', sans-serif;
+	font-family: 'Raleway', sans-serif;"
+/>
